Replace samegroup with samerole
samegroup was introduced in PostgreSQL 7.3[1], but has been obsoleted by
samerole since PostgreSQL 8.1[2].

This patch updates the documentation and scripts to use samerole instead
of samegroup; it could perhaps be improved by giving a different warning
if samegroup is detected. The Mandriva spec file has not been touched,
as again migration from samegroup to samerole needs to be implemented
there.

[1] http://www.postgresql.org/docs/7.3/static/client-authentication.html
[2] http://www.postgresql.org/docs/8.1/static/client-authentication.html